Add basic support for ssh on non-Plan9 platforms
This adds basic support for ssh transport on platforms other than Plan 9.

It isn't very intelligent, it only looks for ~/.ssh/id_rsa for the private
key and if it finds it, uses it.
