+++ git/ssh_other.go
@@ -1,0 +1,18 @@
+// +build !plan9
+// +build !dragonfly
+// +build !openbsd
+// +build !darwin
+// +build !freebsd
+// +build !netbsd
+// +build !solaris
+
+package git
+
+import (
+	"golang.org/x/crypto/ssh"
+)
+
+func getSigners() ([]ssh.Signer, error) {
+	// By default, assume no keys
+	return nil, nil
+}
